backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
pyjwt>=2.9.0,<3.0.0
chromadb>=0.5.0,<1.0.0
pymupdf>=1.24.0,<2.0.0
pandas>=2.2.0,<3.0.0
openai>=1.50.0,<2.0.0
structlog>=24.0.0,<25.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.12,<1.0.0
google-generativeai>=0.8.0,<1.0.0
fpdf2>=2.7.0,<3.0.0
